+++ baseX/markup_search.php
@@ -30,13 +30,14 @@ function noResult($string ) {
  * (C) BaseX Team 2005-12, BSD License
  */
 include("BaseXClient.php");
+
 try {
   
   // create session
   $session = new Session("localhost", 1984, "admin", "admin");
   
   try {
-	if (file_exists("../results/.dirs.dat")) {
+	if (file_exists("../results/.dirs.dat") && basename($_SERVER['HTTP_REFERER']) !== "results.php") {
 		unlink("../results/.dirs.dat");
 	}
 	
@@ -61,28 +62,35 @@ try {
 		$url='http://localhost/colenso/';
   		print '<META HTTP-EQUIV=REFRESH CONTENT="1; '.$url.'">';
 	} else {
-		$query = $session->query("declare namespace tei= 'http://www.tei-c.org/ns/1.0'; ".$input);
-			
-		// loop through all results
-		$i = 0;
-		$iter = 0;
-		while($query->more()) {
-		  $iter++;
-		  if ((strcmp($range, "custom") == 0 && ($iter < $lower || $iter > $upper))) {
-			  //skip
-			$query->next();
-		  } else {
-			//put it one in a xml file
-			$myfile = fopen("../results/search".($i+=1).".xml", "w") or die("Unable to open file!");
-			fwrite($myfile, $query->next());
-			fclose($myfile);
-		  }
+	  $query = $session->query("declare namespace tei= 'http://www.tei-c.org/ns/1.0'; ".$input);
 		  
-		  
+	  // loop through all results
+	  $i = 0;
+	  $iter = 0;
+	  while($query->more()) {
+		$iter++;
+		if ((strcmp($range, "custom") == 0 && ($iter < $lower || $iter > $upper))) {
+			//skip
+		  $query->next();
+		} else {
+		  //put it one in a xml file
+		  $myfile = fopen("../results/search".($i+=1).".xml", "w") or die("Unable to open file!");
+		  fwrite($myfile, $query->next());
+		  fclose($myfile);
 		}
-		$query->info();
-		// close query instance
-		$query->close();
+	  }
+	  $query->info();
+	  // close query instance
+	  $query->close();
+	  
+	  if (file_exists("../results/.dirs.dat") && basename($_SERVER['HTTP_REFERER']) === "results.php") {
+	  copy("../results/.dirs.dat", ".dirs.dat");
+	  } 
+	  if (file_exists(".dirs.dat") && basename($_SERVER['HTTP_REFERER']) !== "results.php") {
+		unlink(".dirs.dat");
+	  }
+	  
+	  
 	}
   } catch (Exception $e) {
     // print exception
